fix(qualifier): reject decimal scores/weights before they hit Int columns
A model emitting 7.5 for a 0-10 score, or a client submitting 50.5/49.5
weights, threw an uncaught write failure after the (paid) Claude call had
already succeeded. Add .int() to both Zod schemas, round again at the
persistence boundary as a backstop, and tell the model explicitly that
scores must be integers. Adds boundary (0/10) and decimal-rejection tests.

+++ src/app/api/jobs/[id]/rubric/route.ts
@@ -77,7 +77,7 @@ export async function POST(_req: Request, { params }: { params: Promise<{ id: st
 const putSchema = z.object({
   criteria: z.array(z.object({
     name: z.string().min(1),
-    weight: z.number().min(0).max(100),
+    weight: z.number().int().min(0).max(100),
     description: z.string(),
   })).min(1),
 });
